WhispAdapt/requirements.txt
datasets
transformers
huggingface_hub
pyyaml
peft
evaluate
scikit-learn
matplotlib
seaborn
pyyaml
accelerate
librosa
